Move FontVariant to it's own File
